Implemented shader load and prism

+++ src/graphic/Fast3D/gfx_metal_shader.cpp
@@ -106,7 +106,7 @@ static const char* shader_item_to_str(uint32_t item, bool with_alpha, bool only_
 
 #undef RAND_NOISE
 
-static void append_formula(char* buf, size_t* len, const uint8_t c[2][4], bool do_single, bool do_multiply, bool do_mix,
+static void append_formula(char* buf, size_t* len, const int c[2][4], bool do_single, bool do_multiply, bool do_mix,
                            bool with_alpha, bool only_alpha, bool opt_alpha, bool first_cycle) {
     if (do_single) {
         append_str(buf, len,
